Updated code cave.

+++ Reverse-Me.cpp
@@ -8,10 +8,15 @@
 class AssemblyCode
 {
 public:
-	byte codeCave[6] =
+	byte codeCave[33] =
 	{
-		0x48, 0x89, 0x54, 0x24, 0x08,	//mov [rsp+8], rdx
-		0xC3,							//ret
+		0x4C, 0x89, 0x44, 0x24, 0x12,	//mov [rsp+0x12], r8
+		0x48, 0x89, 0x54, 0x24, 0x0A,	//mov [rsp+0xa], rdx
+		0x48, 0x89, 0x4C, 0x24, 0x08,	//mov [rsp+0x8], rcx
+		0x48, 0x8B, 0x44, 0x24, 0x08,	//mov rax, [rsp+0x8]
+		0x48, 0x8B, 0x4C, 0x24, 0x0A,   //mov rcx, [rsp+0xa]
+		0x48, 0x89, 0x08,				//mov [rax], rcx
+		0x48, 0x8B, 0x44, 0x24, 0x08    //mov rax, [rsp+0x8]
 	};
 	byte call[23] =
 	{
@@ -76,11 +81,23 @@ int main()
 		assemblyCode.call[i] = *ptr;
 		ptr++;
 	}
-	SIZE_T bytes;
 
-	bool test = WriteProcessMemory(processInformation.hProcess, (LPVOID)writeAddr, assemblyCode.call, 23, &bytes);
-	test = WriteProcessMemory(processInformation.hProcess, (LPVOID)codeCaveAddr, assemblyCode.codeCave, sizeof(assemblyCode.codeCave) / sizeof(assemblyCode.codeCave[0]), &bytes);
-	int lasterror = GetLastError();
+	SIZE_T bytes;
+	if (!WriteProcessMemory(processInformation.hProcess, (LPVOID)writeAddr, assemblyCode.call, 23, &bytes))
+	{
+		printf("Error: %d \n", GetLastError());
+		system("pause");
+		exit(0);
+	};
+	if (!WriteProcessMemory(processInformation.hProcess, (LPVOID)codeCaveAddr, assemblyCode.codeCave, sizeof(assemblyCode.codeCave) / sizeof(assemblyCode.codeCave[0]), &bytes))
+	{
+		printf("Error: %d \n", GetLastError());
+		system("pause");
+		exit(0);
+	};
+
+
+
 
 	nativeMethods.NtResumeProcess(processInformation.hProcess);
 	TerminateProcess(processInformation.hProcess, 0);
